Restructure flake - many changec
New apps:
    Chek app-buchet & catalogue

update: niri-unstable now working!
This was ecause home-manager was established as a stang-alone config,
and also because I didnt know the difference between the syntax of
home-manager packages and nixos system-wide packages. So the answer was
to change syntax and structure.

I also mave a config for niri settings. It's not perfect, I have many
settings commented out, and I haven't converted the window-rules.kvl yet
but it's all more functional.

Pending:
 - optimizing structure
 - imlementing omore shel stuff
 - get some sleep