Organize for theming - configure terminal stuff
Lots of progress on spicetify, ghostty, starship, zsh, yazi, all that
integration. What's untouched is color theming and coordination, as well
as the the shell. Things are running smoothly however. yazi as a
file-chooser is different than dolphin, but has high potential, and much
simpler solutions to malfunctions than trying to configure the kde suite
to work together. Light-weight, rust, terminal-centered, highly
configurable with plugins too. I think it's worth it.

I think it would be good to provide a shortcut for showing yazi keybinds
as well though, since I'm still learning them! and keybinds aren't the
most user friendly. What would be even better is if drag-and-grop were
implemented somehow.