Install spicetify, fix fonts, .configs
Mostly changes in how programs are being configured. Going from
any sort of options for programs to defaulting to typical .config
options, since it's standard and actually relates to any documentation
you find on the web for your program (as opposed to having to interpret
settings through an under-documented nix translation layer). It's more
fragmented, but this flake is already designed to be paired with another
dotfiles repo, or in an assets folder here or something. Not sure yet,
I'll look into some other flakes for common practice.

    - Check the impermanence.nix changes in preserved files to track
    what config files are being created and stowed.
    - I think coming soon will be full dynamic theming with mutagen,
    next, quickshell, then probably LUKS encryption, which will probably
    involve some pretty large commits since I'll have to push dotfiles
    as well before wiping the disks.

    For dotfiles as well, maybe I mentioned this in an earlier commit,
but this is intriguing: https://r.je/evict-your-darlings, and when it's
makes sense, I'll look into this whole process. Perhaps right before the
encryption would be a good time. Pace is picking up, and this desktop
feels much closer to coming out of the "alpha" stage, so to speak.

I gotta update the readme, that'll be a separate commit.