Add support for pianoteq, add libre office
I use a virtual piano controller, and pianoteq is proprietary software
from Modartt with strict licensing rights. Pianoteq binaries are not
going to be included, of course. If you would like to also use pianoteq
in your use of this flake, download pianoteq from Modartt.com with your
own login, then unzip the 7z file to .flake/pkgs/pianoteq/src, and then
enable the package in your user-apps.nix file.

Because this commit includes changes to user-apps.nix, I'll also mention
the addition of the libre-office suite and zathura for lightweight
dynamic pdf reloading (was using that with texlivefull, which is large
and currently disabled.)